+++ src/components/event-rollup-btn.js
@@ -1,0 +1,5 @@
+export const createEventRollupBtnTemplate = () => {
+  return (`<button class="event__rollup-btn" type="button">
+                        <span class="visually-hidden">Open event</span>
+                      </button>`);
+};
