+++ src/components/trip-no-point.js
@@ -1,0 +1,3 @@
+export const createTripNoPointTemplate = () => {
+  return (`<p class="trip-events__msg">Click New Event to create your first point</p>`);
+};
